Refactored FirmwareUpgradeMode into its own file
There's just too much stuff here. I can't keep my head straight.
